Refactor sidebar navigation data flow
